+++ app/assets/javascripts/routers/front/DashboardRouter.js
@@ -139,7 +139,8 @@
      * Init the description's "Read more" button
      */
     _initDescription: function () {
-      document.querySelector('.js-read-more').addEventListener('click', this._openDescriptionModal.bind(this));
+      var readModeBtn = document.querySelector('.js-read-more');
+      if (readModeBtn) readModeBtn.addEventListener('click', this._openDescriptionModal.bind(this));
     },
 
     /**