+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 14 Feb 2017
+- In a site's dataset listing, only the datasets belonging to a site's context are show, even for admins.
 - Fixed bug that was duplicating the routes
 - Added site's contexts and default contexts
 - Fixed an issue where the breadcrumbs wouldn't appear for third-level pages